moved Loading to be a component
